Move tests to data-driven using YAML files #4
This was done by collecting the list of touched files (with a temp
override of get_test_loc() ) then touch empty yaml files and
regenerating the tests. Finally the yaml files have been hand reviewed
for correctness.

All tests with complex expectations have been kept as-is and not
migrated to data-driven tests


Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>